chore: refactor cfd password modal
