+++ packages/cfd/src/Containers/cfd-dashboard.tsx
@@ -21,7 +21,7 @@ import MT5TradeModal from './mt5-trade-modal';
 import CFDDbviOnboarding from './cfd-dbvi-onboarding';
 import CFDDownloadContainer from '../Components/cfd-download-container';
 import CFDPasswordManagerModal from './cfd-password-manager-modal';
-import CFDPasswordModal from './cfd-password-modal';
+import CFDPasswordModal from './cfd-password-modal/cfd-password-modal';
 import CFDServerErrorDialog from './cfd-server-error-dialog';
 import CFDTopUpDemoModal from './cfd-top-up-demo-modal';
 import CFDResetPasswordModal from './cfd-reset-password-modal';
