+++ packages/cfd/src/Containers/index.js
@@ -1,5 +1,5 @@
 import CFDDashboard from './cfd-dashboard';
-import CFDPasswordModal from './cfd-password-modal';
+import CFDPasswordModal from './cfd-password-modal/cfd-password-modal';
 import CFDServerErrorDialog from './cfd-server-error-dialog';
 
 export { CFDPasswordModal };
